requirements.txt
# Dependencias para SekretariaBot
# ================================

# Telegram Bot
python-telegram-bot>=20.0

# Google Calendar API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Gemini 3 Flash Preview (IA)
google-genai

# Utilidades
python-dotenv>=1.0.0
